requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-deriv-api==0.1.6
reactivex==4.0.4
typing_extensions==4.12.2
websocket-client==1.8.0
websockets==10.3
Werkzeug==3.1.3
yarl==1.18.3
